Rename test_Plm.{c,h} to test_lnPlm.{c,h}
